+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -84,6 +84,7 @@ object Constants {
         const val INTELLIJ_PLATFORM_TEST_PLUGIN_LOCAL = "intellijPlatformTestPluginLocal"
         const val INTELLIJ_PLATFORM_TEST_PLUGIN = "intellijPlatformTestPlugin"
         const val INTELLIJ_PLATFORM_TEST_RUNTIME_CLASSPATH = "intellijPlatformTestRuntimeClasspath"
+        const val INTELLIJ_PLATFORM_TEST_RUNTIME_FIX_CLASSPATH = "intellijPlatformTestRuntimeFixClasspath"
         const val INTELLIJ_PLATFORM_TEST_BUNDLED_PLUGINS = "intellijPlatformTestBundledPlugins"
         const val INTELLIJ_PLATFORM_TEST_BUNDLED_MODULES = "intellijPlatformTestBundledModules"
 
@@ -145,7 +146,7 @@ object Constants {
             const val RUNTIME_ELEMENTS = JavaPlugin.RUNTIME_ELEMENTS_CONFIGURATION_NAME
             const val RUNTIME_ONLY = JavaPlugin.RUNTIME_ONLY_CONFIGURATION_NAME
             const val TEST_COMPILE_CLASSPATH = JvmConstants.TEST_COMPILE_CLASSPATH_CONFIGURATION_NAME
-            const val TEST_IMPLEMENTATION = JvmConstants.TEST_IMPLEMENTATION_CONFIGURATION_NAME
+            const val TEST_COMPILE_ONLY = JvmConstants.TEST_COMPILE_ONLY_CONFIGURATION_NAME
             const val TEST_RUNTIME_CLASSPATH = JvmConstants.TEST_RUNTIME_CLASSPATH_CONFIGURATION_NAME
             const val TEST_FIXTURES_COMPILE_ONLY = "testFixturesCompileOnly"
             const val TEST_FIXTURES_COMPILE_CLASSPATH = "testFixturesCompileClasspath"