+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/TestIdeTask.kt
@@ -9,7 +9,9 @@ import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.UntrackedTask
 import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.assign
+import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.named
+import org.jetbrains.intellij.platform.gradle.Constants.Configurations
 import org.jetbrains.intellij.platform.gradle.Constants.Plugin
 import org.jetbrains.intellij.platform.gradle.Constants.Sandbox
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
@@ -87,6 +89,7 @@ abstract class TestIdeTask : Test(), TestableAware, IntelliJPlatformVersionAware
 
             val sourceSets = project.extensions.getByName("sourceSets") as SourceSetContainer
             val runtimeDependencies = sourceSets.getByName(SourceSet.MAIN_SOURCE_SET_NAME).runtimeClasspath
+            val testRuntimeFixConfiguration = project.configurations[Configurations.INTELLIJ_PLATFORM_TEST_RUNTIME_FIX_CLASSPATH]
 
             // The below is needed to simulate the behavior of com.intellij.ide.plugins.cl.PluginClassLoader
             // which is present in the IDE when the plugin is used in "production".
@@ -114,6 +117,7 @@ abstract class TestIdeTask : Test(), TestableAware, IntelliJPlatformVersionAware
             // 4. Test classpath configuration
             // 5. Original classpath without runtime dependencies
             // 6. Test runtime classpath configuration
+            // 7. Test runtime fixes classpath configuration, see: https://youtrack.jetbrains.com/issue/IJPL-180516
             classpath = project.files(
                 instrumentedTestCode,
                 currentPluginLibsProvider,
@@ -121,6 +125,7 @@ abstract class TestIdeTask : Test(), TestableAware, IntelliJPlatformVersionAware
                 sourceTask.intellijPlatformTestClasspathConfiguration,
                 classpath.filter { it !in runtimeDependencies.files },
                 sourceTask.intellijPlatformTestRuntimeClasspathConfiguration,
+                testRuntimeFixConfiguration,
             )
 
             testClassesDirs = instrumentedTestCode + testClassesDirs